fix: f64 time precision — prevent beat/time truncation during long shows
The master time `current_time_double` was f64, but beat position was
derived as f32 (losing ~3 decimal digits after 4+ hours). Generator
context also passed f32 time/beat. This widens the CPU-side precision
chain to f64 throughout, with f32 downcast only at GPU uniform boundaries.

Key changes:
- TempoMapConverter: add seconds_to_beat_f64() for f64 beat derivation
- PlaybackEngine: current_beat is now f64 internally, update_beat_from_time
  uses f64 path through current_time_double, f32 accessor preserved
- TickContext.frame_count: i32 → u64 (prevents overflow at 60fps after 414 days)
- GeneratorContext: time/beat fields are f64, generators cast `as f32`
  only when writing GPU uniform structs
- StatefulState + generator frame_count fields: u32 → u64
- ContentState.current_beat: f64 for UI thread precision
- All UI consumers cast to f32 at the display boundary

+++ crates/manifold-app/src/app.rs
@@ -456,7 +456,7 @@ impl Application {
     fn navigate_cursor(&mut self, direction: manifold_ui::cursor_nav::Direction) {
         use manifold_ui::cursor_nav::{navigate_cursor, NavResult, NavLayerInfo, NavClipInfo};
 
-        let current_beat = self.selection.insert_cursor_beat.unwrap_or(self.content_state.current_beat);
+        let current_beat = self.selection.insert_cursor_beat.unwrap_or(self.content_state.current_beat as f32);
         let active_idx = self.active_layer_id.as_ref()
             .and_then(|id| self.local_project.timeline.find_layer_index_by_id(id));
         let insert_cursor_idx = self.selection.insert_cursor_layer_id.as_ref()
@@ -1815,7 +1815,7 @@ impl ApplicationHandler for Application {
                     self.import_video_files(std::slice::from_ref(&path));
                 } else if crate::project_io::is_supported_midi_extension(&path) {
                     // MIDI files → route through ProjectIOService
-                    let drop_beat = self.content_state.current_beat;
+                    let drop_beat = self.content_state.current_beat as f32;
                     let drop_layer = self.active_layer_id.as_ref()
                         .and_then(|id| self.local_project.timeline.find_layer_index_by_id(id))
                         .unwrap_or(0) as i32;
