+++ crates/manifold-renderer/src/generators/tesseract.rs
@@ -86,7 +86,7 @@ impl Generator for TesseractGenerator {
         let window = if ctx.param_count > WINDOW as u32 { ctx.params[WINDOW] } else { 0.1 };
         let scale = if ctx.param_count > SCALE as u32 { ctx.params[SCALE] } else { 1.0 };
 
-        let t = ctx.time;
+        let t = ctx.time as f32;
         let angle_xy = t * rot_xy;
         let angle_zw = t * rot_zw;
         let angle_xw = t * rot_xw;
@@ -119,7 +119,7 @@ impl Generator for TesseractGenerator {
             gpu, target,
             positions, instances, num_edges,
             edge_half_thick, dot_half_thick,
-            ctx.beat, "Tesseract", ctx.width, ctx.height,
+            ctx.beat as f32, "Tesseract", ctx.width, ctx.height,
         );
         self.helper.anim_progress
     }
